archive/experiments/lstm/requirements.txt
torch
pandas
scikit-learn
matplotlib
requests
python-dotenv